Move covid estimates panel tooltips to separate module, as that content is expected to grow/go through some refactors
