+++ client/src/panels/panel_declarations/covid/covid_estimates.yaml
@@ -82,49 +82,3 @@ covid_estimates_stat:
 covid_estimates_vote:
   en: Voted Authorities
   fr: Autorisations votées
-
-covid_mains_2021_note:
-  transform: [markdown]
-  en: |
-    The 22.7 billion identified in the Main Estimates for COVID-19 includes approximately $0.6 billion that is outside the scope
-    of Canada’s COVID-19 Economic Response Plan (ERP).
-  fr: |
-    Les 22,7 milliards au titre de la COVID-19 figurant dans le budget principal des dépenses incluent un montant d'environ 0,6 milliard de dollars qui est hors la portée 
-    du plan d'intervention économique.
-covid_supps_a_2021_note:
-  transform: [markdown]
-  en: |
-    The $20.16 billion identified in the Supplementary Estimates (A) for COVID-19 includes approximately $6.02 billion that has been reprofiled from 2020-21.  
-
-    This amount was originally approved in 2020-21, but has subsequently been reprofiled to 2021-22. The voted authorities for 2020-21 will be reported as
-    frozen allotments in the 2021 Public Accounts.
-  fr: |
-    Les 20,16 milliards au titre de la COVID-19 figurant dans le budget supplémentaire des dépenses (A) incluent un montant d'environ 6,02 milliard de
-    dollars qui ont été reportés de 2020-2021.  
-
-    Ce montant a été initialement approuvé en 2020-2021, mais a ensuite été reporté à 2021-2022. Les autorisations votées pour 2020-2021 seront inscrites
-    en tant qu’affectations bloquées dans les Comptes publics 2021.
-
-
-covid_estimates_COV082_2020_tooltip:
-  transform: [handlebars,markdown]
-  en: |
-    Includes $5 billion for vaccines that was approved in the 2020-21 Supplementary Estimates (B), but has subsequently been reprofiled to 2021-22.
-    These voted authorities were reported as frozen allotments in the 2020-21 Supplementary Estimates (C) supplemental information
-    {{ext_link "online" "https://www.canada.ca/en/treasury-board-secretariat/services/planned-government-spending/supplementary-estimates/supplementary-estimates-c-2020-21/frozen-allotments-voted-authorities.html"}}.
-  fr: |
-    Comprend les cinq milliards de dollars pour les vaccins qui ont été approuvés dans le budget supplémentaire des dépenses (B) 2020-2021, mais qui ont ensuite été reportés à 2021-2022.
-    Ces autorisations votées ont été inscrites comme affectations bloquées dans les renseignements supplémentaires
-    {{ext_link "en ligne" "https://www.canada.ca/fr/secretariat-conseil-tresor/services/depenses-prevues/budgets-supplementaires/budget-supplementaire-depenses-c-2020-2021/affectations-bloquees-autorisations-votees.html"}}
-    relatifs au Budget supplémentaire des dépenses (C) 2020-2021.
-covid_estimates_COV115_2020_tooltip:
-  transform: [handlebars,markdown]
-  en: |
-    Includes $500 million for PPE that was approved in the 2020-21 Supplementary Estimates (B), but has subsequently been reprofiled to 2021-22.
-    These voted authorities were reported as frozen allotments in the 2020-21 Supplementary Estimates (C) supplemental information
-    {{ext_link "online" "https://www.canada.ca/en/treasury-board-secretariat/services/planned-government-spending/supplementary-estimates/supplementary-estimates-c-2020-21/frozen-allotments-voted-authorities.html"}}.
-  fr: |
-    Comprend 500 millions de dollars pour les EPI qui ont été approuvés dans le budget supplémentaire des dépenses (B) 2020-2021, mais qui ont ensuite été reportés à 2021-2022.
-    Ces autorisations votées ont été inscrites comme affectations bloquées dans les renseignements supplémentaires
-    {{ext_link "en ligne" "https://www.canada.ca/fr/secretariat-conseil-tresor/services/depenses-prevues/budgets-supplementaires/budget-supplementaire-depenses-c-2020-2021/affectations-bloquees-autorisations-votees.html"}}
-    relatifs au Budget supplémentaire des dépenses (C) 2020-2021.
